@@ -1,6 +1,6 @@
 {
   "name": "@ihk-gfi/lux-components-update",
-  "version": "11.14.0",
+  "version": "13.0.0",
   "description": "Schematics für die Aktualisierung von LUX-Applikationen",
   "scripts": {
     "build": "tsc -p tsconfig.json",
@@ -29,26 +29,23 @@
     "url": "https://github.com/IHK-GfI/lux-components-update/issues"
   },
   "schematics": "./src/collection.json",
-  "devDependencies": {
-    "prettier": "2.0.5"
-  },
   "dependencies": {
-    "@angular-devkit/core": "11.2.13",
-    "@angular-devkit/schematics": "11.2.13",
-    "@angular-devkit/schematics-cli": "^0.1102.13",
-    "@angular/cli": "11.2.13",
-    "@types/jasmine": "3.5.11",
-    "@types/node": "13.7.7",
-    "@types/semver": "7.3.6",
-    "chalk": "4.1.1",
+    "@angular-devkit/core": "13.3.4",
+    "@angular-devkit/schematics": "13.3.4",
+    "@angular-devkit/schematics-cli": "13.3.4",
+    "@angular/cli": "13.3.4",
+    "@types/jasmine": "4.0.3",
+    "@types/node": "16.11.33",
+    "@types/semver": "7.3.9",
+    "chalk": "4.1.2",
     "cheerio": "1.0.0-rc.9",
     "htmlparser2": "6.1.0",
-    "jasmine": "3.7.0",
+    "jasmine": "4.1.0",
     "jsonc-parser": "3.0.0",
-    "semver": "7.3.5",
-    "typescript": "4.2.4"
+    "semver": "7.3.7",
+    "typescript": "4.5.5"
   },
-  "optionalDependencies": {
-    "fsevents": "2.3.2"
+  "devDependencies": {
+    "prettier": "2.5.1"
   }
 }
